requirements.txt
greenlet>=0.4.5
eventlet>=0.17.2
msgpack-python>=0.4.6
